backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
google-generativeai>=0.8
google-genai>=0.1
pydantic>=2.0
pytest>=8.0
httpx>=0.27
bcrypt>=4.0
pyjwt>=2.8
redis>=5.0.0
email-validator
